Stub method and failing tests for get first index tuple of
